Mayor changes across the client and the server
