+++ voting-client/npm-debug.log
@@ -1,0 +1,45 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/bin/nodejs', '/usr/bin/npm', 'start' ]
+2 info using npm@3.8.6
+3 info using node@v5.12.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle voting-client@0.0.0~prestart: voting-client@0.0.0
+6 silly lifecycle voting-client@0.0.0~prestart: no script for prestart, continuing
+7 info lifecycle voting-client@0.0.0~start: voting-client@0.0.0
+8 verbose lifecycle voting-client@0.0.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle voting-client@0.0.0~start: PATH: /usr/lib/node_modules/npm/bin/node-gyp-bin:/home/vagrant/Code/FullStackRedux/voting-client/node_modules/.bin:/usr/bin:/home/vagrant/.composer/vendor/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games
+10 verbose lifecycle voting-client@0.0.0~start: CWD: /home/vagrant/Code/FullStackRedux/voting-client
+11 silly lifecycle voting-client@0.0.0~start: Args: [ '-c', 'node ./bin/www' ]
+12 silly lifecycle voting-client@0.0.0~start: Returned: code: 1  signal: null
+13 info lifecycle voting-client@0.0.0~start: Failed to exec start script
+14 verbose stack Error: voting-client@0.0.0 start: `node ./bin/www`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/usr/lib/node_modules/npm/lib/utils/lifecycle.js:239:16)
+14 verbose stack     at emitTwo (events.js:100:13)
+14 verbose stack     at EventEmitter.emit (events.js:185:7)
+14 verbose stack     at ChildProcess.<anonymous> (/usr/lib/node_modules/npm/lib/utils/spawn.js:24:14)
+14 verbose stack     at emitTwo (events.js:100:13)
+14 verbose stack     at ChildProcess.emit (events.js:185:7)
+14 verbose stack     at maybeClose (internal/child_process.js:850:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:215:5)
+15 verbose pkgid voting-client@0.0.0
+16 verbose cwd /home/vagrant/Code/FullStackRedux/voting-client
+17 error Linux 4.4.0-22-generic
+18 error argv "/usr/bin/nodejs" "/usr/bin/npm" "start"
+19 error node v5.12.0
+20 error npm  v3.8.6
+21 error code ELIFECYCLE
+22 error voting-client@0.0.0 start: `node ./bin/www`
+22 error Exit status 1
+23 error Failed at the voting-client@0.0.0 start script 'node ./bin/www'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the voting-client package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     node ./bin/www
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs voting-client
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls voting-client
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
