+++ voting-client/package.json
@@ -16,6 +16,7 @@
     "ejs": "~2.5.2",
     "express": "~4.14.0",
     "jquery": "^3.1.1",
+    "materialize-css": "^0.97.8",
     "morgan": "~1.7.0",
     "redux": "^3.6.0",
     "redux-logger": "^2.7.4",
