+++ voting-client/src/js/index.js
@@ -1,1 +1,3 @@
 require('./bootstrap');
+
+require('./libs/handleVote');
