+++ voting-client/src/js/rootActionTypes.js
@@ -1,0 +1,10 @@
+var SET_STATE = 'SET_STATE';
+var NEXT = 'NEXT';
+
+exports.SET_STATE;
+exports.NEXT;
+
+module.exports = {
+  SET_STATE: SET_STATE,
+  NEXT: NEXT,
+};
