+++ voting-client/src/js/vote/index.js
