--- voting-client/src/js/votes/actionTypes.js
@@ -1,13 +1,0 @@
-var SET_STATE = 'votes/SET_STATE';
-var VOTE = 'votes/VOTE';
-var NEXT = 'votes/NEXT';
-
-exports.SET_STATE;
-exports.VOTE;
-exports.NEXT;
-
-module.exports = {
-  SET_STATE: SET_STATE,
-  VOTE: VOTE,
-  NEXT: NEXT,
-};
