+++ voting-client/views/index.ejs
@@ -1,12 +1,35 @@
 <!DOCTYPE html>
 <html>
   <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
     <title><%= title %></title>
+
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/5.0.0/normalize.min.css">
+    <link rel='stylesheet' href='/materialize/css/materialize.min.css' />
     <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
   <body>
-    <h1><%= title %></h1>
-    <p>Welcome to <%= title %></p>
+    <h1 class="center">
+      VOTA POR TU PELÍCULA FAVORITA!
+    </h1>
+
+    <div class="row center">
+      <div class="col s3 offset-s3">
+        <button class="waves-effect waves-light btn tile js-voting-button" data-entry="opcion1">
+          Opción 1
+        </button>
+      </div>
+
+      <div class="col s3">
+        <button class="waves-effect waves-light btn tile js-voting-button" data-entry="opcion1">
+          Opción 2
+        </button>
+      </div>
+    </div>
+
     <script src="/javascripts/app.bundle.js"></script>
   </body>
 </html>
