+++ voting-server/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "redux": "^3.6.0",
     "redux-node-logger": "^0.1.0",
-    "reselect": "^2.5.4",
     "socket.io": "^1.7.1"
   }
 }
