+++ voting-server/src/rootActionTypes.js
@@ -1,0 +1,7 @@
+var NEXT = 'NEXT';
+
+exports.NEXT;
+
+module.exports = {
+  NEXT: NEXT,
+};
