+++ voting-server/src/server.js
@@ -4,11 +4,17 @@ module.exports = function startServer(store) {
   var io = new Server().attach(8090);
 
   store.subscribe(function() {
-    return io.emit('state', store.getState());
+    var clientState = Object.assign({}, store.getState());
+    delete clientState.entries;
+
+    return io.emit('state', clientState);
   });
 
   io.on('connection', function(socket) {
-    socket.emit('state', store.getState());
+    var clientState = Object.assign({}, store.getState());
+    delete clientState.entries;
+
+    socket.emit('state', clientState);
     socket.on('action', store.dispatch.bind(store));
   });
 }
