+++ voting-server/src/vote/actionTypes.js
@@ -1,0 +1,7 @@
+var VOTE = 'vote/VOTE';
+
+exports.VOTE;
+
+module.exports = {
+  VOTE: VOTE,
+};
