+++ voting-server/src/vote/index.js
