--- voting-server/src/votes/actionTypes.js
@@ -1,10 +1,0 @@
-var VOTE = 'votes/VOTE';
-var NEXT = 'votes/NEXT';
-
-exports.VOTE;
-exports.NEXT;
-
-module.exports = {
-  VOTE: VOTE,
-  NEXT: NEXT,
-};
